Fix conflicting type arguments error
This patch fixes the conflicting type arguments error when generating
bindings for generic types that use R or P as type arguments, described
at: https://github.com/igloo-project/bindgen/issues/3

The changed behavior will look through a list of preferred type argument
names for both the root and path type arguments, and will pick the first
available one. In case the preferred list of type argument names is
exhausted, it will generate one with the first preferred type argument
name, combined with a number, so that the newly generated name is not
conflicting with any of the type argument names that are already used by
the type for which the binding is generated. The preferred name for the
bindings are:

 - R, ROOT, ROOT_TYPE, or the first available from R_1, R_2, etc.
 - P, PATH, PATH_TYPE, or the first available from P_1, P_2, etc.