Fix problem handling annotated types on type use - continued
The previous commit that supposed to fix the problem with annotated type
use under javac solved only some of the possible cases. Upon further
use, I faced another problem when an annotated type use was combined
with a type that was outside of bindgen's scope. In that case a
GenericObjectBinding was generated for the type and the problem
described in https://github.com/igloo-project/bindgen/issues/8 came back
in a slightly different form. I decided I'll try to do a better job this
time, so I decided to replace all invocations of TypeMirror.toString
with Util.getTypeName. Hopefully this will solve the issue completely.

I also added a JUnit test for the two cases I uncovered up to this
point.