fix getType() method on bindings
In 3987449dad7b4598d6b1375b3211ab917b01c79e, getType() is broken for
some field bindings.

getType() was previously implemented on each generated binding. This was
changed to provide type at \*BindingPath construction time. getType()
value is now stored in bindingType attribute.

But MethodPropertyGenerator.java do not always build a new
\*BindingPath(...) with a type parameter, as it should.

This patch now always passes a type argument if it is available, except
for Array (not sure about it should or should not for this case, so I do
not modify this case).